Polishing #1902
Add since tags. Tweak Javadoc.

Original pull request: #1980.

+++ src/main/java/io/lettuce/core/AbstractRedisAsyncCommands.java
@@ -1607,7 +1607,7 @@ public abstract class AbstractRedisAsyncCommands<K, V> implements RedisAclAsyncC
     }
 
     @Override
-    public RedisFuture<Long> sintercard(int limit, K... keys) {
+    public RedisFuture<Long> sintercard(long limit, K... keys) {
         return dispatch(commandBuilder.sintercard(limit, keys));
     }
