+++ src/main/java/io/lettuce/core/api/async/RedisSetAsyncCommands.java
@@ -18,7 +18,11 @@ package io.lettuce.core.api.async;
 import java.util.List;
 import java.util.Set;
 
-import io.lettuce.core.*;
+import io.lettuce.core.RedisFuture;
+import io.lettuce.core.ScanArgs;
+import io.lettuce.core.ScanCursor;
+import io.lettuce.core.StreamScanCursor;
+import io.lettuce.core.ValueScanCursor;
 import io.lettuce.core.output.ValueStreamingChannel;
 
 /**
@@ -95,24 +99,26 @@ public interface RedisSetAsyncCommands<K, V> {
     RedisFuture<Long> sinter(ValueStreamingChannel<V> channel, K... keys);
 
     /**
-     * This command works exactly like sinter but instead of returning the result set, it returns just the cardinality of the
-     * result.
+     * This command works exactly like {@link #sinter(java.lang.Object[])} but instead of returning the result set, it returns
+     * just the cardinality of the result.
      *
-     * @param keys the key.
+     * @param keys the keys.
      * @return The cardinality of the set which would result from the intersection of all the given sets.
+     * @since 6.2
      */
     RedisFuture<Long> sintercard(K... keys);
 
     /**
-     * This command works exactly like sinter but instead of returning the result set, it returns just the cardinality of the
-     * result.
+     * This command works exactly like {@link #sinter(java.lang.Object[])} but instead of returning the result set, it returns
+     * just the cardinality of the result.
      *
      * @param limit If the intersection cardinality reaches limit partway through the computation, the algorithm will exit and
      *        yield limit as the cardinality.
-     * @param keys the key.
+     * @param keys the keys.
      * @return The cardinality of the set which would result from the intersection of all the given sets.
+     * @since 6.2
      */
-    RedisFuture<Long> sintercard(int limit, K... keys);
+    RedisFuture<Long> sintercard(long limit, K... keys);
 
     /**
      * Intersect multiple sets and store the resulting set in a key.
